fast-ci new
in fast-ci contr subroutines, iprim and jprim are swapped; This is to
resolve the issue with wring rotational energies.
The openmp is now working!
I had to change me_terms to me_term and do openmp for each term.
Now the vibrational part is OK, but the rotational energies are still
wrong.