fast-ci: print out sizes and reducing allocation sizes of field_me
fast-ci: print out sizes and reducing allocation sizes of field_me.
The memory for the Nclass-me objects is too large for CH4, a reporting
is introduced. Beside the actual number of elements, not maximal is
saved.
The part where the check for sparse-parameter is done removed. It needs
to go to the basis set section.