fat-ci and sparse: removed an empty object
in sparse: it can happen that an object has zero elements, for example
external for fittings. Now it should not be referenced  at all. Thee
objects are also deallocated.
Now for the refinement and three fields need to be initialized just so
that the size of the field could be checked.