Fixed segfault
This error only occured with gfortran due to the array being accessed
within an `any` call. I presume the intel compiler assumes this to be
false. However, that behaviour negates the whole point of the code which
checks the values that are intended to be in the array. The offending
code is when processing the DIAGONALIZER input block in `FLReadInput` in
`fields.f90`.

The array is necessarily allocated when the symmetry is defined so the
DIAGONALIZER block must come after this. I've added a check to make sure
the symmetry block runs first but this breaks the input files for the
benchmark I'm using and probably breaks others.