Include Intel compiler and MPI in CI (#36)
* Start running with Intel compiler

* Use correct compiler command

I don't like this but it seems there's no way to conditionally
set the environment variable. See also
https://github.com/actions/runner/issues/409
Will possibly change to conditional steps later.

* Remember Intel-related parameters

The scripts set the PATH (and maybe other variables?) but those are
not preserved across steps. Since the build step is becoming more
distinct between the two compilers, let's try splitting it in two
conditional ones.

* Try caching Intel installation

* Look for Intel libraries before tests

* Try to cache whole directory

* Allow running tests with MPI

* Fix syntax

* Install MPI compilers when needed

* Try to force use of ifort with MPI

By default, the Intel version of mpif90 uses gfortran. We can
either switch to using mpiifort instead, or try to control the
underlying compiler this way. See also, for example:
https://www.hpc.cineca.it/center_news/important-use-intel-mpi-wrappers-mpif90-mpicc-mpicxx

* Make sure cache names don't clash with/without MPI

* Update variables so MKL and BLACS can be found

* Fix MKL installation

* Show some more info and build faster on CI

* Only use one MPI process on CI for debugging

See if the error still happens when testing.

* Avoid reading from stdin

* Build faster on CI with gfortran

* Avoid segmentation fault with quick gfortran build

This reverts commit b8413c6593b4a1b7344239801683ccb5ce499539.

* Replicate execution on cluster temporarily

Adding the mpiio option may be required for now but will eventually
be removed.

* Don't test file_intensity with MPI

It's currently failing when run with MPI and > 1 processes
(not just on GitHub Actions, also on CSD3).

* Simplify USE_MPI in Makefile and CI

Now behaving similar to the other Makefile variables. This also
lets us make the GitHub Actions job a bit simpler.

* Don't install MKL twice when using MPI, clean up