Avoid polymorphism problems with Intel compiler
To use an unlimited-polymorphic value as argument in a subroutine,
the corresponding dummy argument must also be declared as unlimited
polymorphic. This isn't the case with the MPI_File_Write routines,
so we have to explicitly check the type each time.
The gfortran compiler is happy to ignore this (or perhaps it looks
up the right subroutine to use automatically) - this error only
appears with the Intel compiler.