Use concrete types so tests pass with Intel
At this point, the MPI tests pass when using 2 or 4 processes.
The second test (writing a column-distributed array) fails when
using only 1 process only due to a dimension mismatch between
read and write (which is maybe to be expected, as I don't think
that code is called in TROVE when comm_size is 1).