New KEO for XY2 was added with all necessary components
MLkinetic_compact_xy2_bisect_EKE_rigid_singular_rho is a KEO as an expansion in term of all three modes with a proper treatment of singularity using csc_.
This is accompanied by new coordinates r-rho-z-abs for absolute ranges of the modes.
This has been made to work with ME_sinrho_Legendre_k both for RIGID and non-RIGID. Some changes and fixes to fields and me_bnd were applied.